feat(intro): So i got the intro to work and this is the prototype and i also change the image of Bitbox
