+++ bitbox/app/_layout.tsx
@@ -20,22 +20,15 @@ export default function RootLayout() {
 
 
     return (
-        <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
-            <Stack>
-                <Stack.Screen name="index" options={{ headerShown: false }} />
-                <Stack.Screen name="login" />
-            </Stack>
-            <StatusBar style="auto" />
+        <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
+                 <Stack>
+                     <Stack.Screen name="index" options={{ headerShown: false }} />
+                     <Stack.Screen name="login" />
+                 </Stack>
+                 <StatusBar style="auto" />
         </ThemeProvider>
     );
 }
 
-
-//
-// <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
-//     <Stack>
-//         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-//         <Stack.Screen name="+not-found" />
-//     </Stack>
-//     <StatusBar style="auto" />
-// </ThemeProvider>
+// name="(tabs)"
+// name="+not-found"
