+++ bitbox/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.1.0",
+    "@lottiefiles/dotlottie-react": "^0.13.5",
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.9",
@@ -31,6 +32,7 @@
     "expo-system-ui": "~5.0.7",
     "expo-video": "~2.1.9",
     "expo-web-browser": "~14.1.6",
+    "lottie-react-native": "^7.2.2",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "0.79.2",
